// src/web/src/api/generated/model/productImportPreviewDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Zalihe.Web | v1
 * OpenAPI spec version: 1.0.0
 */
import type { ProductToCreateDto } from './productToCreateDto';
import type { ProductToLinkDto } from './productToLinkDto';
import type { SkippedProductDto } from './skippedProductDto';

export interface ProductImportPreviewDto {
  linkedCount: number;
  toLink: ProductToLinkDto[];
  toCreate: ProductToCreateDto[];
  skipped: SkippedProductDto[];
}